feat: updated to have a sticky header
