+++ packages/nextjs/components/Header/NavItems.ts
@@ -1,0 +1,10 @@
+export const NAV_ITEMS = [
+  {
+    href: "/products",
+    label: "Products",
+  },
+  {
+    href: "/categories",
+    label: "Categories",
+  },
+];
