+++ packages/nextjs/components/Layout.tsx
@@ -1,7 +1,7 @@
 import Head from "next/head";
 import React from "react";
 import { Footer } from "./Footer/Footer";
-import { Header } from "./Header";
+import { Header } from "./Header/Header";
 
 interface Props {
   children: React.ReactNode;
@@ -14,8 +14,12 @@ export const Layout: React.FC<Props> = ({ children }) => {
         <title>Formidable Boulangerie</title>
       </Head>
       <div className="h-screen flex flex-col">
-        <Header />
-        <main className="w-full">{children}</main>
+        <div className="w-full fixed z-10 bg-yellow">
+          <Header />
+        </div>
+        <div className="w-full sm:pt-[142px] pt-[98px]">
+          <main className="w-full">{children}</main>
+        </div>
         <Footer />
       </div>
     </>
